Feat: Des nouveaux trucs

+++ src/controllers/user-routes.js
@@ -1,4 +1,5 @@
 const express = require('express');
+const jwt = require('jsonwebtoken')
 const router = express.Router();
 const userRepository = require('../models/user-repository');
 
@@ -17,16 +18,37 @@ router.get('/:firstName', (req, res) => {
 });
 
 router.post('/', (req, res) => {
+  let token = req.headers.authorization.split(' ')[1]
+  token = jwt.verify(token, 'Wong Xi Fang Su Ha')
+  if(token.user.role != 'ADMIN') {
+    res.status(403).end()
+    throw new Error('You are not an admin, acces not allowed');
+  }
+
   userRepository.createUser(req.body);
   res.status(201).end();
 });
 
 router.put('/:id', (req, res) => {
+  let token = req.headers.authorization.split(' ')[1]
+  token = jwt.verify(token, 'Wong Xi Fang Su Ha')
+  if(token.user.role != 'ADMIN') {
+    res.status(403).end()
+    throw new Error('You are not an admin, acces not allowed');
+  }
+
   userRepository.updateUser(req.params.id, req.body);
   res.status(204).end();
 });
 
 router.delete('/:id', (req, res) => {
+  let token = req.headers.authorization.split(' ')[1]
+  token = jwt.verify(token, 'Wong Xi Fang Su Ha')
+  if(token.user.role != 'ADMIN') {
+    res.status(403).end()
+    throw new Error('You are not an admin, acces not allowed');
+  }
+
   userRepository.deleteUser(req.params.id);
   res.status(204).end();
 });
